🚚 api: remove controller to certification/enrolment context
